refactor: extract shared buildBlossomEdges, runBlossom, and normaliseGames
buildBlossomEdges and runBlossom move to pairing-helpers.ts — identical
across Dubov, Burstein, and Lim except for the system tag and criteria
array. normaliseGames moves to utilities.ts — was duplicated in Burstein
and Lim. net ~180 lines removed.
